RDoc-2389 Configure Revisions Operation - Part 2
